# [Intel MKL] CNMS performance optimization by using priority queue