Replace vector with priority queue to improve the performance in CNMS kernel